fix: allow disabling encryption and block access to encryption settings when not decrypted
